==> Scripts/03_treinamento_do_modelo_de_decisao.ipynb <==
##### Configurações Iniciais

import os
import pandas as pd

from Modules.operation_analiser import OperationLoader, OperationAnalyzer
from Modules.forecast_loader import ForecastLoader
from Modules.variables_info import variables_info
from Modules.model_metrics import ModelPlotter
from Modules.scenario_classifier import RainScenarioClassifier

# Parâmetros gerais
lat_porto = -25.500511
lon_porto = -48.519331
forecast_path = '../Database/forecast_data'
paradas_path = '../Database/Parada_2024_chuva.xlsx'
timebox = ('202405', '202412')

variaveis = variables_info['variaveis']
legendas = variables_info['legendas']
legendas_dict = dict(zip(variaveis, legendas))

# Repositorios de Saída
base_output = '../Output/Analise_03'
os.makedirs('../Output/Analise_03', exist_ok=True)
for subfolder in ['Modelos', 'Tables', 'Imagens', 'Imagens/Modelos', 'Imagens/Modelos/Modelo_01', 'Imagens/Modelos/Modelo_02', 'Imagens/Metricas/Modelo_01', 'Imagens/Metricas/Modelo_02', 'Modelos/Modelo_01', 'Modelos/Modelo_02']:
    os.makedirs(os.path.join(base_output, subfolder), exist_ok=True)

##### Importação dos Dados para Analise

forecast = ForecastLoader(forecast_path, lat_target=lat_porto, lon_target=lon_porto)
df_forecast = forecast.load_forecasts(timebox, drop_duplicates=True)

operation_loader = OperationLoader(paradas_path)
df_paradas = operation_loader.load()

analyzer = OperationAnalyzer(df_paradas, df_forecast)
df_unificado = analyzer.marcar_paradas()

melhores = pd.read_csv(f'../Output/Analise_02/Tables/melhores_thresholds.csv')

#### Treinamento do Modelo Supervisionado para Classificação das Recomendações
##### Baseado nas Metricas

cenarios = {
    'conservador': melhores.sort_values('recall', ascending=False).head(3),
    'moderado': melhores.sort_values('f1_score', ascending=False).head(3),
    'arrojado': melhores.sort_values('precision', ascending=False).head(3),
}

modelo = RainScenarioClassifier(df_unificado, cenarios, base_output, legendas_dict)
modelo.aplicar_classificacao(modelo_nome="Modelo_01")

##### Analise dos Modelos

# Carrega e prepara o DataFrame
df_metricas = pd.read_csv(f'{base_output}/Tables/metricas_modelo_01_por_cenario.csv')

plotter = ModelPlotter(df_metricas)

# 1. Comparação geral (sem filtro)
plotter.plotar_metricas(
    titulo='Desempenho Geral dos Modelos - Previsão de Paradas e Não-Paradas',
    output_path=f'{base_output}/Imagens/Modelos/Modelo_01/comparacao_geral_modelos.png'
)

# 2. Comparação por cenário - Classe: Parada
plotter.plotar_metricas(
    titulo='Comparação por Cenário - Classe: Parada',
    filtro_classe=1,
    col='cenario',
    output_path=f'{base_output}/Imagens/Modelos/Modelo_01/comparacao_por_cenario_parada.png'
)

# 3. Comparação por cenário - Classe: Não-Parada
plotter.plotar_metricas(
    titulo='Comparação por Cenário - Classe: Não-Parada',
    filtro_classe=0,
    col='cenario',
    output_path=f'{base_output}/Imagens/Modelos/Modelo_01/comparacao_por_cenario_nao_parada.png'
)

# 4. Comparação por cenário e por classe
plotter.plotar_metricas(
    titulo='Comparação por Cenário e Classe',
    col='cenario',
    row='classe',
    output_path=f'{base_output}/Imagens/Modelos/Modelo_01/comparacao_por_cenario_e_classe.png'
)

#### Treinamento do Modelo Supervisionado para Classificação das Recomendações
##### Combinação Variaveis (Estatistica + Metricas)

cenarios = {
    'conservador': [('precipitation_probability_55km', 15.0)],
    'moderado': [('precipitation_probability_35km', 25.0), ('lwe_precipitation_smooth_rate_maximum_35km', 0.5)],
    'arrojado': [('precipitation_probability_20km', 30.0), ('lwe_precipitation_smooth_rate_maximum_20km', 0.4)]
}

modelo = RainScenarioClassifier(df_unificado, cenarios, base_output, legendas_dict)
modelo.aplicar_classificacao(modelo_nome="Modelo_02")

##### Analise dos Modelos

# Carrega e prepara o DataFrame
df_metricas = pd.read_csv(f'{base_output}/Tables/metricas_modelo_02_por_cenario.csv')

# Instancia o plotador com o DataFrame carregado
plotter = ModelPlotter(df_metricas)

# 1. Comparação geral (sem filtro)
plotter.plotar_metricas(
    titulo='Desempenho Geral dos Modelos - Previsão de Paradas e Não-Paradas',
    output_path=f'{base_output}/Imagens/Modelos/Modelo_02/comparacao_geral_modelos.png'
)

# 2. Comparação por cenário - Classe: Parada
plotter.plotar_metricas(
    titulo='Comparação por Cenário - Classe: Parada',
    filtro_classe=1,
    col='cenario',
    output_path=f'{base_output}/Imagens/Modelos/Modelo_02/comparacao_por_cenario_parada.png'
)

# 3. Comparação por cenário - Classe: Não-Parada
plotter.plotar_metricas(
    titulo='Comparação por Cenário - Classe: Não-Parada',
    filtro_classe=0,
    col='cenario',
    output_path=f'{base_output}/Imagens/Modelos/Modelo_02/comparacao_por_cenario_nao_parada.png'
)

# 4. Comparação por cenário e por classe
plotter.plotar_metricas(
    titulo='Comparação por Cenário e Classe',
    col='cenario',
    row='classe',
    output_path=f'{base_output}/Imagens/Modelos/Modelo_02/comparacao_por_cenario_e_classe.png'
)
